requirements.txt
uvicorn
fastapi
speedtest-cli